Add files via upload
This C++ code demonstrates the use of a simple class `Test` with a user-defined constructor and destructor. The constructor outputs a message when an object of the class is created, while the destructor outputs a message when the object is destroyed. The program creates an instance of the `Test` class, showing the execution of both the constructor and destructor.