Add files via upload
This program demonstrates function overloading by defining two add() functions: one for adding integers and another for adding doubles. It uses function overloading to handle different data types with the same function name. In main(), the integer addition (10 + 20) and double addition (5.5 + 7.8) are performed. The results are printed to the console. Output: Sum of integers 10 and 20: 30 and Sum of doubles 5.5 and 7.8: 13.3.